radiation_ie_mvp_B/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
lxml>=5.0
beautifulsoup4>=4.12
joblib>=1.3
scikit-learn>=1.3
